web_api/requirements.txt
asgiref==3.4.1
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
gunicorn==20.1.0
pytz==2021.3
sqlparse==0.4.2
